refactor: abstract viewall and currency utils
