+++ src/bot/convo_handlers/Settleup/utils/reports.py
@@ -8,8 +8,12 @@ from telegram.error import BadRequest
 from telegram.ext import ContextTypes
 
 from src.bot.convo_handlers.Settleup.utils.general import get_settleup_details
-from src.lib.currencies.config import ALL_CURRENCY_CODES, CURRENCY_SHORTHAND_MAPPING
-from src.lib.currencies.utils import convert, read_cached_exchange_rates
+from src.lib.currencies.config import ALL_CURRENCY_CODES
+from src.lib.currencies.utils import (
+    convert,
+    get_shorthand_currency,
+    read_cached_exchange_rates,
+)
 from src.lib.splizy_repo.model import ExpenseRow
 
 matplotlib.use("Agg")
@@ -66,7 +70,7 @@ def _build_metadata_lines(
     report_generated_at: datetime,
 ) -> list[str]:
     settle_currency = settleup_currency.upper()
-    settle_shorthand = CURRENCY_SHORTHAND_MAPPING.get(settle_currency, settle_currency)
+    settle_shorthand = get_shorthand_currency(settle_currency)
     settle_desc = ALL_CURRENCY_CODES.get(settle_currency, settle_currency)
 
     lines = [
